+++ spec/www/spec/db-tx-string-test.js
@@ -1472,15 +1472,15 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        // TBD KNOWN ENCODING ISSUE with emojis and other 4-byte UTF-8 characters
+        // ENCODING ISSUE NOW FIXED with emojis and other 4-byte UTF-8 characters
         // on default Android evcore NDK implementation
         // ref:
         // - litehelpers/Cordova-sqlite-evcore-extbuild-free#44
         // - litehelpers/Cordova-sqlite-storage#564
         // - litehelpers/Cordova-sqlite-evcore-extbuild-free#7
 
-        it(suiteName + 'string HEX value test with UTF-8 4-byte Gothic bairkan 𐌱 (U+10331) [XXX ENCODING BUG REPRODUCED on default Android SQLite3 XXX on default Android SQLite3 NDK build (using Android-evcore-native-driver); default sqlite HEX encoding: UTF-6le on Windows & Android 4.1-4.3 (WebKit) Web SQL, UTF-8 otherwise]', function(done) {
-          // ENCODING BUG REPRODUCED for 4-byte UTF-8 characters
+        it(suiteName + 'string HEX value test with UTF-8 4-byte Gothic bairkan 𐌱 (U+10331) [ENCODING ISSUE NOW FIXED on default Android evcore NDK implementation; default sqlite HEX encoding: UTF-6le on Windows & Android 4.1-4.3 (WebKit) Web SQL, UTF-8 otherwise]', function(done) {
+          // ENCODING ISSUE NOW FIXED for 4-byte UTF-8 characters
           // on default Android database access implementation
           // (Android-sqlite-connector with Android-sqlite-ext-native-driver,
           // using NDK) on Android pre-6.0
@@ -1500,8 +1500,6 @@ var mytests = function() {
                 expect(rs1.rows.item(0).myresult).toBe('400000D831DF'); // (UTF-16le)
               else if (!isWebSql && !isWindows && isAndroid && !isImpl2 && (/Android [4-5]/.test(navigator.userAgent)))
                 expect(rs1.rows.item(0).myresult).toBe('40EDA080EDBCB1'); // ENCODING ISSUE REPRODUCED on default Android NDK implementation on Android 4.x/5.x
-              else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
-                expect(rs1.rows.item(0).myresult).toBe('403F'); // ENCODING ISSUE REPRODUCED on default Android NDK post-5.x
               else
                 expect(rs1.rows.item(0).myresult).toBe('40F0908CB1'); // (UTF-8)
 
@@ -1513,8 +1511,6 @@ var mytests = function() {
                   expect(rs2.rows.item(0).myresult).toBe('400000D831DF'); // (UTF-16le)
                 else if (!isWebSql && !isWindows && isAndroid && !isImpl2 && (/Android [4-5]/.test(navigator.userAgent)))
                   expect(rs2.rows.item(0).myresult).toBe('40EDA080EDBCB1'); // ENCODING ISSUE REPRODUCED on default Android NDK pre-6.0
-                else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
-                  expect(rs2.rows.item(0).myresult).toBe('403F'); // ENCODING ISSUE REPRODUCED on default Android NDK post-5.x
                 else
                   expect(rs2.rows.item(0).myresult).toBe('40F0908CB1'); // (UTF-8)
 
@@ -1533,7 +1529,7 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + 'string manipulation test with UTF-8 4-byte Gothic bairkan 𐌱 (U+10331) - TBD ENCODING ISSUE REPRODUCED on Android post-5.x (default evcore NDK implementation)', function(done) {
+        it(suiteName + 'string manipulation test with UTF-8 4-byte Gothic bairkan 𐌱 (U+10331) - ENCODING ISSUE NOW FIXED on default Android SQLite3 evcore NDK build', function(done) {
           // ref:
           // - litehelpers/Cordova-sqlite-evcore-extbuild-free#44
           // - litehelpers/Cordova-sqlite-storage#564
@@ -1546,19 +1542,13 @@ var mytests = function() {
               expect(rs1).toBeDefined();
               expect(rs1.rows).toBeDefined();
               expect(rs1.rows.length).toBe(1);
-              if (!isWebSql && isAndroid && !isImpl2 && !(/Android [4-5]/.test(navigator.userAgent)))
-                expect(rs1.rows.item(0).myresult).toBe('A?'); // ENCODING ISSUE on Android post-5.x
-              else
-                expect(rs1.rows.item(0).myresult).toBe('A𐌱');
+              expect(rs1.rows.item(0).myresult).toBe('A𐌱');
 
               tx.executeSql("SELECT UPPER('a𐌱') AS myresult", null, function(ignored, rs2) {
                 expect(rs2).toBeDefined();
                 expect(rs2.rows).toBeDefined();
                 expect(rs2.rows.length).toBe(1);
-                if (!isWebSql && isAndroid && !isImpl2 && !(/Android [4-5]/.test(navigator.userAgent)))
-                  expect(rs2.rows.item(0).myresult).toBe('A?'); // ENCODING ISSUE on Android post-5.x
-                else
-                  expect(rs2.rows.item(0).myresult).toBe('A𐌱');
+                expect(rs2.rows.item(0).myresult).toBe('A𐌱');
 
                 // Close (plugin only) & finish:
                 (isWebSql) ? done() : db.close(done, done);
@@ -1573,11 +1563,15 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + "SELECT LOWER(X'41F0908CB1') - RETURNS '\\uF041\\u8C90' ('\uF041\u8C90') UTF-16le on Android 4.1-4.3 (WebKit) Web SQL & Windows, 'a?' on Android with default evcore NDK provider (encoding issue), UTF-8 'a\\uD800\\uDF31' ('a\uD800\uDF31') otherwise", function(done) {
+        it(suiteName + "SELECT LOWER(X'41F0908CB1') - RETURNS '\\uF041\\u8C90' ('\uF041\u8C90') UTF-16le on Android 4.1-4.3 (WebKit) Web SQL & Windows, UTF-8 'a\\uD800\\uDF31' ('a\uD800\uDF31') otherwise - ENCODING ISSUE NOW FIXED on default Android evcore NDK implementation", function(done) {
+          // ENCODING ISSUE NOW FIXED with emojis and other 4-byte UTF-8 characters
+          // on default Android evcore NDK implementation
           // ref:
           // - litehelpers/Cordova-sqlite-evcore-extbuild-free#44
           // - litehelpers/Cordova-sqlite-storage#564
           // - litehelpers/Cordova-sqlite-evcore-extbuild-free#7
+          if (!isWebSql && isAndroid && !isImpl2 && /Android 5/.test(navigator.userAgent)) pending('POSSIBLE CRASH on Android 5 with default Android evcore NDK implementation');
+
           var db = openDatabase('SELECT-LOWER-X-41F0908CB1-test.db');
 
           db.transaction(function(tx) {
@@ -1587,8 +1581,6 @@ var mytests = function() {
               expect(rs.rows.length).toBe(1);
               if (isWindows || (isWebSql && isAndroid && /Android 4.[1-3]/.test(navigator.userAgent)))
                 expect(rs.rows.item(0).lowertext).toBe('\uF041\u8C90'); // (UTF-16le)
-              else if (!isWebSql && isAndroid && !isImpl2)
-                expect(rs.rows.item(0).lowertext).toBe('a?'); // ENCODING ISSUE on Android (default evcore NDK)
               else
                 expect(rs.rows.item(0).lowertext).toBe('a\uD800\uDF31'); // 'a𐌱' (UTF-8)
 
@@ -1637,7 +1629,7 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + 'Inline emoji string manipulation test: SELECT UPPER("a\\uD83D\\uDE03.") [\\u1F603 SMILING FACE (MOUTH OPEN)] - ENCODING ISSUE REPRODUCED on Android post-5.x', function(done) {
+        it(suiteName + 'Inline emoji string manipulation test: SELECT UPPER("a\\uD83D\\uDE03.") [\\u1F603 SMILING FACE (MOUTH OPEN)] - ENCODING ISSUE NOW FIXED on default Android evcore NDK implementation', function(done) {
           // ref:
           // - litehelpers/Cordova-sqlite-evcore-extbuild-free#44
           // - litehelpers/Cordova-sqlite-storage#564
@@ -1652,10 +1644,7 @@ var mytests = function() {
               expect(rs).toBeDefined();
               expect(rs.rows).toBeDefined();
               expect(rs.rows.length).toBe(1);
-              if (!isWebSql && isAndroid && !isImpl2 && !(/Android [4-5]/.test(navigator.userAgent)))
-                expect(rs.rows.item(0).uppertext).toBe('A?.'); // ENCODING ISSUE on Android post-5.x
-              else
-                expect(rs.rows.item(0).uppertext).toBe('A\uD83D\uDE03.');
+              expect(rs.rows.item(0).uppertext).toBe('A\uD83D\uDE03.');
 
               // Close (plugin only) & finish:
               (isWebSql) ? done() : db.close(done, done);
@@ -1669,7 +1658,7 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + 'emoji string argument value manipulation test - ENCODING ISSUE REPRODUCED on Android post-5.x (default evcore NDK implementation)', function(done) {
+        it(suiteName + 'emoji string argument value manipulation test - ENCODING ISSUE NOW FIXED on default Android evcore NDK implementation', function(done) {
           // ref:
           // - litehelpers/Cordova-sqlite-evcore-extbuild-free#44
           // - litehelpers/Cordova-sqlite-storage#564
@@ -1685,10 +1674,7 @@ var mytests = function() {
               expect(rs.rows).toBeDefined();
               expect(rs.rows.length).toBe(1);
 
-              if (!isWebSql && isAndroid && !isImpl2 && !(/Android [4-5]/.test(navigator.userAgent)))
-                expect(rs.rows.item(0).uppertext).toBe('A?.'); // ENCODING ISSUE REPRODUCED on Android post-5.x
-              else
-                expect(rs.rows.item(0).uppertext).toBe('A\uD83D\uDE03.');
+              expect(rs.rows.item(0).uppertext).toBe('A\uD83D\uDE03.');
 
               // Close (plugin only) & finish:
               (isWebSql) ? done() : db.close(done, done);
@@ -1702,9 +1688,9 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + 'emoji HEX test: SELECT HEX("@\\uD83D\\uDE03!") [\\u1F603 SMILING FACE (MOUTH OPEN)] [TBD ENCODING ISSUE REPRODUCED on Android (default evcore NDK implementation); default sqlite HEX encoding: UTF-6le on Windows & Android 4.1-4.3 (WebKit) Web SQL, UTF-8 otherwise]', function(done) {
-          // ENCODING ISSUE REPRODUCED for emojis and other 4-byte UTF-8
-          // characters on default Android evcore NDK implementation
+        it(suiteName + 'emoji HEX test: SELECT HEX("@\\uD83D\\uDE03!") [\\u1F603 SMILING FACE (MOUTH OPEN)] [default sqlite HEX encoding: UTF-6le on Windows & Android 4.1-4.3 (WebKit) Web SQL, UTF-8 otherwise]', function(done) {
+          // ENCODING ISSUE NOW FIXED with emojis and other 4-byte UTF-8 characters
+          // on default Android evcore NDK implementation
           // ref:
           // - litehelpers/Cordova-sqlite-evcore-extbuild-free#44
           // - litehelpers/Cordova-sqlite-storage#564
@@ -1724,8 +1710,6 @@ var mytests = function() {
                 expect(rs.rows.item(0).hexvalue).toBe('40003DD803DE2100'); // (UTF-16le)
               else if (!isWebSql && !isWindows && isAndroid && !isImpl2 && /Android [4-5]/.test(navigator.userAgent))
                 expect(rs.rows.item(0).hexvalue).toBe('40EDA0BDEDB88321'); // ENCODING ISSUE REPRODUCED on Android pre-6.0
-              else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
-                expect(rs.rows.item(0).hexvalue).toBe('403F21'); // ENCODING ISSUE REPRODUCED on Android post-5.x
               else
                 expect(rs.rows.item(0).hexvalue).toBe('40F09F988321'); // (UTF-8)
 
@@ -1738,8 +1722,6 @@ var mytests = function() {
                   expect(rs2.rows.item(0).hexvalue).toBe('40003DD803DE2100'); // (UTF-16le)
                 else if (!isWebSql && !isWindows && isAndroid && !isImpl2 && /Android [4-5]/.test(navigator.userAgent))
                   expect(rs2.rows.item(0).hexvalue).toBe('40EDA0BDEDB88321'); // ENCODING ISSUE REPRODUCED on Android pre-6.0
-                else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
-                  expect(rs2.rows.item(0).hexvalue).toBe('403F21'); // ENCODING ISSUE REPRODUCED on Android post-5.x
                 else
                   expect(rs2.rows.item(0).hexvalue).toBe('40F09F988321'); // (UTF-8)
 
@@ -1758,11 +1740,13 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + "Inline BLOB with emoji string manipulation test: SELECT LOWER(X'41F09F9883') - RETURNS '\\uF041\\u989F' ('\uF041\u989F') UTF-16le on Android 4.1-4.3 (WebKit) Web SQL & Windows, 'a?' on Android (default evcore NDK provider), UTF-8 'a\\uD83D\\uDE03' ('a\uD83D\uDE03') with U+1F603 SMILING FACE (MOUTH OPEN) otherwise", function(done) {
+        it(suiteName + "Inline BLOB with emoji string manipulation test: SELECT LOWER(X'41F09F9883') - RETURNS '\\uF041\\u989F' ('\uF041\u989F') UTF-16le on Android 4.1-4.3 (WebKit) Web SQL & Windows,  UTF-8 'a\\uD83D\\uDE03' ('a\uD83D\uDE03') with U+1F603 SMILING FACE (MOUTH OPEN) otherwise", function(done) {
           // ref:
           // - litehelpers/Cordova-sqlite-evcore-extbuild-free#44
           // - litehelpers/Cordova-sqlite-storage#564
           // - litehelpers/Cordova-sqlite-evcore-extbuild-free#7
+          if (!isWebSql && isAndroid && !isImpl2 && /Android 5/.test(navigator.userAgent)) pending('POSSIBLE CRASH on Android 5 with default Android evcore NDK implementation');
+
           var db = openDatabase('SELECT-LOWER-X-41F09F9883-test.db');
           expect(db).toBeDefined();
 
@@ -1775,8 +1759,6 @@ var mytests = function() {
               expect(rs.rows.length).toBe(1);
               if (isWindows || (isWebSql && isAndroid && /Android 4.[1-3]/.test(navigator.userAgent)))
                 expect(rs.rows.item(0).lowertext).toBe('\uF041\u989F'); // (UTF-16le)
-              else if (!isWebSql && isAndroid && !isImpl2)
-                expect(rs.rows.item(0).lowertext).toBe('a?'); // ENCODING ISSUE REPRODUCED on Android
               else
                 expect(rs.rows.item(0).lowertext).toBe('a\uD83D\uDE03');
 
@@ -2093,7 +2075,7 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + 'HEX value of string with 25 emojis [TBD ENCODING ISSUE reproduced on Android (default evcore NDK provider) on Android post-5.x (non-standard encoding on Android 4.x/5.x); default sqlite HEX encoding: UTF-6le on Windows & Android 4.1-4.3 (WebKit) Web SQL, UTF-8 otherwise]', function(done) {
+        it(suiteName + 'HEX value of string with 25 emojis [non-standard encoding on Android 4.x/5.x; default sqlite HEX encoding: UTF-6le on Windows & Android 4.1-4.3 (WebKit) Web SQL, UTF-8 otherwise]', function(done) {
           // ref:
           // - litehelpers/Cordova-sqlite-evcore-extbuild-free#44
           // - litehelpers/Cordova-sqlite-storage#564
@@ -2133,9 +2115,6 @@ var mytests = function() {
                   '40EDA0BDEDB881EDA0BDEDB882EDA0BDEDB883EDA0BDEDB884EDA0BDEDB885' +
                   '40EDA0BDEDB881EDA0BDEDB882EDA0BDEDB883EDA0BDEDB884EDA0BDEDB885' +
                   '40EDA0BDEDB881EDA0BDEDB882EDA0BDEDB883EDA0BDEDB884EDA0BDEDB885');
-              else if (!isWebSql && isAndroid && !isImpl2)
-                expect(resultRow1.hexValue).toBe(
-                  '403F3F3F3F3F403F3F3F3F3F403F3F3F3F3F403F3F3F3F3F403F3F3F3F3F');
               else
                 expect(resultRow1.hexValue).toBe(
                   '40F09F9881F09F9882F09F9883F09F9884F09F9885' +
@@ -2154,7 +2133,7 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + 'UPPER value of string with 25 emojis - ENCODING ISSUE REPRODUCED on Android post-5.x (default evcore NDK implementation)', function(done) {
+        it(suiteName + 'UPPER value of string with 25 emojis', function(done) {
           // ref:
           // - litehelpers/Cordova-sqlite-evcore-extbuild-free#44
           // - litehelpers/Cordova-sqlite-storage#564
@@ -2175,10 +2154,7 @@ var mytests = function() {
               var resultRow1 = rs1.rows.item(0);
               expect(resultRow1).toBeDefined();
               expect(resultRow1.upperText).toBeDefined();
-              if (!isWebSql && isAndroid && !isImpl2 && !(/Android [4-5]/.test(navigator.userAgent)))
-                expect(resultRow1.upperText).toBe('A?????A?????A?????A?????A?????');
-              else
-                expect(resultRow1.upperText).toBe(
+              expect(resultRow1.upperText).toBe(
                   'A\uD83D\uDE01\uD83D\uDE02\uD83D\uDE03\uD83D\uDE04\uD83D\uDE05' +
                   'A\uD83D\uDE01\uD83D\uDE02\uD83D\uDE03\uD83D\uDE04\uD83D\uDE05' +
                   'A\uD83D\uDE01\uD83D\uDE02\uD83D\uDE03\uD83D\uDE04\uD83D\uDE05' +