+++ CHANGES.md
@@ -2,6 +2,13 @@
 
 # cordova-sqlite-evcore-extbuild-free 0.13.0-dev
 
+- update libs from cordova-sqlite-evcore-free-dependencies@0.10.1
+  - BREAKING: drop Android pre-5.1 support
+  - SQLite 3.35.5 update
+  - fix for emojis and other 4-byte UTF-8 characters on Android
+  - update affected test cases
+  - update README.md
+
 ## cordova-sqlite-evcore-feat-android-db-location 0.8.0
 
 - refactor: update androidDatabaseLocation error messages